requirements.txt
google-adk
google-generativeai
python-dotenv
pypdf
fpdf2
